+++ NEWS.md
@@ -73,6 +73,7 @@ Misc
 - Replace Base-64 codec to fix erroneous failures and add a raw encoder
 - atheme.conf.example: document SET NOPASSWORD module
 - Services will no longer begin a new database unless passed the -b option
+- Port contrib/ns_waitreg to modules/nickserv/waitreg
 
 Build System
 ------------